add user validation to expense categories

+++ src/expense-categories/expense-categories.service.ts
@@ -6,19 +6,23 @@ import { PrismaService } from '../prisma.service';
 export class ExpenseCategoriesService {
   constructor(private prisma: PrismaService) {}
 
-  async addNewCategory(category: ExpenseCategoryDto) {
+  async addNewCategory(category: ExpenseCategoryDto, user: any) {
     category.name = category.name.toUpperCase().trim();
 
-    const candidate = await this.prisma.expenseCategory.findUnique({ where: { name: category.name } });
+    const candidateCategoryId = await this.ifCategoryExistsReturnsItsId(category.name, user.sub);
 
-    if (candidate) {
-      throw new HttpException('Expense category already exists', 400);
+    console.log(`addNewCategory: ${candidateCategoryId}`);
+
+    if (candidateCategoryId !== 0) {
+      throw new HttpException('Category already exists', 400);
     }
 
-    console.log(`NAME: ${category.name}`);
     const createdCategory = await this.prisma.expenseCategory.create({
       data: {
         name: category.name,
+        user: {
+          connect: { id: user.sub },
+        },
       },
     });
 
@@ -27,11 +31,15 @@ export class ExpenseCategoriesService {
     return createdCategory;
   }
 
-  async removeCategory(id: number) {
+  async removeCategory(id: number, user: any) {
     const candidate = await this.prisma.expenseCategory.findUnique({ where: { id: id } });
 
     if (!candidate) {
       throw new HttpException('No objects found', 400);
+    }
+
+    if (candidate.user_id !== user.sub) {
+      throw new HttpException('Access not allowed', 401);
     }
 
     const removed = await this.prisma.expenseCategory.delete({ where: { id: id } });
@@ -39,5 +47,21 @@ export class ExpenseCategoriesService {
     console.log(`Remove category with id ${removed.id}`);
 
     return removed;
+  }
+
+  async ifCategoryExistsReturnsItsId(categoryName: string, user_id: number): Promise<number> {
+    categoryName = categoryName.toUpperCase().trim();
+
+    const candidates = await this.prisma.expenseCategory.findMany({ where: { user_id: user_id } });
+
+    let candidateId = 0;
+
+    candidates.forEach(function (candidate) {
+      if (candidate.name == categoryName) {
+        candidateId = candidate.id;
+      }
+    });
+
+    return candidateId;
   }
 }
